backend/requirements.txt
aiomysql
argon2-cffi
cryptography
fastapi
httpx
orjson
pyyaml
ua-parser
user-agents
uvicorn[standard]
PyJWT
SQLAlchemy
